+++ src/Core/BoundedContexts/Headquarters/WB.Core.BoundedContexts.Headquarters/Views/Reposts/SurveyStatistics/Data/GetReportCategoricalPivotReportItem.cs
@@ -2,8 +2,8 @@
 {
     public class GetReportCategoricalPivotReportItem
     {
-        public int RowValue { get; set; }
-        public int ColValue { get; set; }
+        public long RowValue { get; set; }
+        public long ColValue { get; set; }
         public long Count { get; set; }
     }
 }
